modular xsb storage pooler
